+++ src/util.jsx
@@ -1,0 +1,3 @@
+export const If = function(props) {
+    return props.test ? props.children : null;
+}