+++ template/src/build.js
@@ -14,12 +14,11 @@ if (fs.pathExistsSync(dist_folder)) {
 fs.mkdirsSync(dist_folder);
 
 // run `webpack --mode production`
-const config = require("../webpack.config.js");
+const config = require("../webpack.config.js")({production: true});
 config.output = {
     path: dist_folder,
     filename: 'adventure.js'
 };
-config.mode = 'production';
 
 webpack(config, (err, stats) => {
     // errors